+++ src/main/java/br/com/ifpb/ads/bookifyapi/repository/BookRepository.java
@@ -4,5 +4,8 @@ package br.com.ifpb.ads.bookifyapi.repository;
 import br.com.ifpb.ads.bookifyapi.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<Book, Integer> {
+import java.util.List;
+
+public interface BookRepository extends JpaRepository<Book, Long> {
+
 }
